Reorganizing creatorScreen into more files
Also deleting unnecessary files
